+++ qoqo/src/devices/square_lattice.rs
@@ -94,8 +94,8 @@ impl SquareLatticeDeviceWrapper {
     /// Set gate time of all two-qubit gates of specific type
     ///
     /// Args:
-    ///     gate[str]: The hqslang name of the two-qubit-gate.
-    ///     gate_time[f64]: Gate time for the given gate, valid for all qubits in the device.
+    ///     gate (str): The hqslang name of the two-qubit-gate.
+    ///     gate_time (float): Gate time for the given gate, valid for all qubits in the device.
     ///
     /// Returns:
     ///     A qoqo Device with updated gate times.
@@ -113,8 +113,8 @@ impl SquareLatticeDeviceWrapper {
     /// Set gate time of all single-qubit gates of specific type
     ///
     /// Args:
-    ///     gate[str]: The hqslang name of the single-qubit-gate.
-    ///     gate_time[f64]: New gate time.
+    ///     gate (str): The hqslang name of the single-qubit-gate.
+    ///     gate_time (float): New gate time.
     ///
     /// Returns:
     ///     A qoqo Device with updated gate times.
@@ -156,7 +156,7 @@ impl SquareLatticeDeviceWrapper {
     /// Adds qubit damping to noise rates.
     ///
     /// Args:
-    ///     daming (float): The damping rates.
+    ///     damping (float): The damping rates.
     ///
     /// Returns:
     ///     SquareLatticeDevice