+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 This changelog track changes to the qoqo project starting at version v0.5.0
 
-## Unreleased
+## 1.13.1
 
-### Fixed in Unreleased
+### Added in 1.13.1
 
+* Added the doc_generator feature to qoqo that will create .pyi used for python IDE hints during cargo build.
+* Change version mismatch error message to be clearer.
 * `PragmaSetStateVector`, `PragmaSetDensityMatrix` and `PragmaGeneralNoise` interface casting errors when handling arrays/matrices
 
 ## 1.13.0